chore: separate params from options
