+++ packages/client-fetch/src/index.ts
@@ -4,8 +4,10 @@ export type {
   ClientOptions,
   Config,
   CreateClientConfig,
+  OmitNever,
   Options,
   OptionsLegacyParser,
+  Params,
   RequestOptions,
   RequestResult,
   TDataShape,
