+++ packages/openapi-ts/package.json
@@ -87,6 +87,7 @@
     "typescript": "^5.5.3"
   },
   "devDependencies": {
+    "@hey-api/client-core": "workspace:*",
     "@types/cross-spawn": "6.0.6",
     "@types/express": "4.17.21",
     "axios": "1.8.2",
