+++ packages/openapi-ts/src/plugins/@hey-api/sdk/constants.ts
@@ -1,3 +1,5 @@
 export const nuxtTypeComposable = 'TComposable';
 export const nuxtTypeDefault = 'DefaultT';
 export const nuxtTypeResponse = 'ResT';
+
+export const sdkId = 'sdk';
