+++ packages/openapi-ts/src/plugins/@hey-api/sdk/response.ts
@@ -1,0 +1,53 @@
+/**
+ * Infers `responseType` value from provided response content type. This is
+ * an adapted version of `getParseAs()` from the Fetch API client.
+ *
+ * From Axios documentation:
+ * `responseType` indicates the type of data that the server will respond with
+ * options are: 'arraybuffer', 'document', 'json', 'text', 'stream'
+ * browser only: 'blob'
+ */
+export const getResponseType = (
+  contentType: string | null | undefined,
+):
+  | 'arraybuffer'
+  | 'blob'
+  | 'document'
+  | 'json'
+  | 'stream'
+  | 'text'
+  | undefined => {
+  if (!contentType) {
+    return;
+  }
+
+  const cleanContent = contentType.split(';')[0]?.trim();
+
+  if (!cleanContent) {
+    return;
+  }
+
+  if (
+    cleanContent.startsWith('application/json') ||
+    cleanContent.endsWith('+json')
+  ) {
+    return 'json';
+  }
+
+  // Axios does not handle form data out of the box
+  // if (cleanContent === 'multipart/form-data') {
+  //   return 'formData';
+  // }
+
+  if (
+    ['application/', 'audio/', 'image/', 'video/'].some((type) =>
+      cleanContent.startsWith(type),
+    )
+  ) {
+    return 'blob';
+  }
+
+  if (cleanContent.startsWith('text/')) {
+    return 'text';
+  }
+};
