+++ packages/openapi-ts/src/plugins/@hey-api/sdk/typeOptions.ts
@@ -1,13 +1,10 @@
-import ts from 'typescript';
-
 import { compiler } from '../../../compiler';
 import { clientModulePath } from '../../../generate/client';
 import type { FileImportResult } from '../../../generate/files';
 import type { IR } from '../../../ir/types';
 import type { Plugin } from '../../types';
 import { getClientPlugin } from '../client-core/utils';
-import { nuxtTypeDefault, nuxtTypeResponse } from './constants';
-import { sdkId } from './plugin';
+import { nuxtTypeDefault, nuxtTypeResponse, sdkId } from './constants';
 import type { Config } from './types';
 
 export const createTypeOptions = ({
@@ -132,42 +129,4 @@ export const createTypeOptions = ({
   });
 
   file.add(typeOptions);
-};
-
-export const createTypeOmitNever = ({ context }: { context: IR.Context }) => {
-  const file = context.file({ id: sdkId })!;
-
-  const neverType = compiler.keywordTypeNode({ keyword: 'never' });
-  const kType = compiler.typeReferenceNode({ typeName: 'K' });
-  const tType = compiler.typeReferenceNode({ typeName: 'T' });
-  const kOfTType = compiler.indexedAccessTypeNode({
-    indexType: kType,
-    objectType: tType,
-  });
-
-  const omitNeverTypeAlias = compiler.typeAliasDeclaration({
-    exportType: true,
-    name: 'OmitNever',
-    type: ts.factory.createMappedTypeNode(
-      undefined,
-      ts.factory.createTypeParameterDeclaration(
-        undefined,
-        'K',
-        ts.factory.createTypeOperatorNode(ts.SyntaxKind.KeyOfKeyword, tType),
-        undefined,
-      ),
-      ts.factory.createConditionalTypeNode(
-        kOfTType,
-        neverType,
-        neverType,
-        kType,
-      ),
-      undefined,
-      kOfTType,
-      undefined,
-    ),
-    typeParameters: [{ name: 'T' }],
-  });
-
-  file.add(omitNeverTypeAlias);
 };
